HSEARCH-4505 Implement a scenario that is more likely to fail in ElasticsearchIndexSchemaManagerUpdateCustomMappingIT#illegalUpdate
Switching from '_source: true' to '_source: false' is obviously
something that Elasticsearch could support, and from Elasticsearch 8.1
it does (though not on 8.0 and below).

Switching from '_source: false' to '_source: true', on the other hand,
is not possible because the source for already-indexed document is no
longer available, so this should fail even on Elasticsearch 8.1.