HSEARCH-4691 Backport HSEARCH-4484 Align byte-buddy dependency on Hibernate ORM even in Spring tests
Spring uses an older version of byte-buddy that won't work with recent
versions of Mockito, so we have to explicitly manage that dependency.
But we don't want to diverge from the version used by Hibernate ORM,
hence the limited dependency management (only for spring-related ITs)
and hacks to get maven-enforcer-plugin to check for dependency
alignment.