HSEARCH-4571 Fix flaky test OutOfSyncIndexIT
If we just wait for "at least one" document to be indexed, then it's
possible that some documents haven't been indexed yet, leading to the
test failing when it checks that all documents are indexed.

Better make indexing synchronous, as we do in almost every other test
(including tests of the Library showcase).