HSEARCH-4654 Add an entry to the failure report instead of throwing an exception when reaching the failure limit in RootFailureCollector
The rendering of the failure report in that exception used to lead to
a deadlock, because we would:

* lock on a child to add a failure
* check on the root whether there are more failures than the limit<
* lock on the root to access the children
* try to lock on each child one after the other,
  to render their failures and include that in the "failure limit
  reached" exception message

So we can't render the failure report for that exception. But if that
exception doesn't include the failure report in its message, then we run
the risk of aborting Hibernate Search bootstrap without

There were other potential problems with throwing an exception, as the
risk of ending up adding that exception to the failure report because of
surrounding try blocks; which would not alter the failure report, of
course, but would still be logged, and would throw another exception
about exceeding the failure limit. This could result in very verbose
logs.

So, the easiest solution is probably to just add a simple message to the
error report, and to proceed with whatever we were doing, until we
finally end up checking the error report and abort, displaying the error
report along with a mention that there were too many failures to report.