HSEARCH-4708 Allow considering empty objects as equivalent to a missing object when comparing indexed documents in tests
This will be mostly useful because Hibernate ORM might create empty
embeddables, which will only make automatic indexing tests more
complicated since we would have to take those into account when
comparing documents.

If we can consider empty embeddables as equivalent to missing objects in
indexed documents, then assertions stay as simple as they used to be,
in particular in tests that are not related to embeddables.