HSEARCH-4727 Fix definition of "payload" columns in outbox events with DB2
This is a backwards-incompatible change for DB2 users, but the schema
was non-fonctional for those users to begin with, so that's no big deal.

This is also potentially a breaking change for users of some exotic
databases whose dialect had the same problem as DB2's, but those
databases are not tested nor advertised as supported anyway.